add chunk reader.

+++ Ssq/Chunk.cs
@@ -4,7 +4,7 @@ namespace Ssq
 {
     public class Chunk
     {
-        public int Offset { get; set; }
+        public long Offset { get; set; }
         public ChunkHeader Header { get; set; }
         public uint[] TimeOffsets { get; set; } = Array.Empty<uint>();
         public int[] Tempo_TFPS_Config { get; set; } = Array.Empty<int>();
